third_party/WebKit/LayoutTests/fast/dom/custom/document-register-on-create-callback-expected.txt

document.registerElement() in a createdCallback should upgrade elements that are in the same processing queue.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS createdElements is ["elemA0", "elemB0"]
PASS createdElements is ["elemA0", "elemB0", "elemB1"]
PASS successfullyParsed is true

TEST COMPLETE
